Do not cache component peers
As of this commit, we derive component peers upon each key event.

The reason for this change, is that cached peers may go obsolete.
This can be seen in Reaper when using Keyboard in an AUv3. When
closing the editor, the AudioProcessorEditor's destructor is not
actually invoked. When reopening the editor, the peer, or host
window, is not the same instance as it was before closing it.

Since peer derivation is not expensive, and since we're not doing
it on the audio thread, we can afford to do this.