Support key repeat synthesis for some keys
AUv3 for some reason does not propagate key repeats into the key
event monitor for some keys. So far, the _/- and +/= keys have
been confirmed to suffer from this.
Maybe it's because AUv3 runs in a sandbox. AUv2 does not have
this problem.

VMPC2000XL, the project that is the reason for this JUCE module,
by default has said keys configured for turning the DATA wheel.
This happens to be the one action where key repeat is vital.

Desperate times lead to desperate measures. We hardcode the key
codes for these two keys, and synthesize key repeats on a
juce::Timer.

There should be no breaking changes. Synthesis can be enabled
by passing `true` as the second argument for the
`KeyboardFactory::instance(...)` method, which has a default
value of `false`.